+++ domain/task.go
@@ -4,8 +4,8 @@ import "time"
 
 // Task is a single unit of work under a story.
 type Task struct {
-	ID        uint      `json:"id"`
-	StoryID   uint      `json:"storyId"`
+	ID        uint64    `json:"id"`
+	StoryID   uint64    `json:"storyId"`
 	Name      string    `json:"name"`
 	Status    string    `json:"status"`
 	CreatedAt time.Time `json:"createdAt"`
