+++ keeper/story.go
@@ -11,7 +11,7 @@ type StoryKeeper interface {
 // StoryReader reads stories
 type StoryReader interface {
 	// GetStory reads a single story
-	GetStory(id uint) (domain.Story, error)
+	GetStory(id uint64) (domain.Story, error)
 	// GetStories reads a page of stories
 	GetStories(limit, offset int) []domain.Story
 }
@@ -21,7 +21,7 @@ type StoryWriter interface {
 	// CreateStory creates a new story
 	CreateStory(title string) (domain.Story, error)
 	// UpdateStory updates an existing story
-	UpdateStory(id uint, title string) (domain.Story, error)
+	UpdateStory(id uint64, title string) (domain.Story, error)
 	// DeleteStory deletes an existing story
-	DeleteStory(id uint) (int64, error)
+	DeleteStory(id uint64) (int64, error)
 }
