+++ keeper/task.go
@@ -11,17 +11,17 @@ type TaskKeeper interface {
 // TaskReader reads tasks
 type TaskReader interface {
 	// GetTask reads a single task
-	GetTask(id uint) (domain.Task, error)
+	GetTask(id uint64) (domain.Task, error)
 	// GetTasks reads a page of tasks for a story
-	GetTasks(storyID uint, limit, offset int) []domain.Task
+	GetTasks(storyID uint64, limit, offset int) []domain.Task
 }
 
 // TaskWriter writes tasks
 type TaskWriter interface {
 	// CreateTask creates a new story
-	CreateTask(storyID uint, name string) (domain.Task, error)
+	CreateTask(storyID uint64, name string) (domain.Task, error)
 	// UpdateTask updates an existing story
-	UpdateTask(id uint, name, status string) (domain.Task, error)
+	UpdateTask(id uint64, name, status string) (domain.Task, error)
 	// DeleteTask deletes an existing story
-	DeleteTask(id uint) (int64, error)
+	DeleteTask(id uint64) (int64, error)
 }
